applying reshape for mmstack tiff images
